requirements.txt
PyYAML==5.3.1
numpy==1.19.4
pandas==1.1.4
torch==1.7.0
tqdm==4.46.0
scikit-image==0.17.2
matplotlib==3.3.3
Pillow==8.0.1
#sklearn==0.0
scikit-learn==0.23.2
torchvision==0.8.1
setuptools==46.4.0.post20200518
tensorboardX==2.1
tensorflow==2.3.1
scipy==1.5.4